Shop page branch

+++ app/src/main/java/com/mhmdawad/superest/presentation/main/adapter/CartAdapter.kt
@@ -56,7 +56,7 @@ constructor() : RecyclerView.Adapter<CartAdapter.CartViewHolder>() {
     ) {
         val quantity = text.toString()
         if (quantity.isNotEmpty()) {
-            val quantityNumber = quantity.toDouble()
+            val quantityNumber = quantity.toDouble().toInt()
             if (quantityNumber > 0) {
                 val product = purchasedProductsList.single { it.id == productModel.id }
                 product.quantity = quantityNumber
